Use InfixOperatorExpr
